// Fase-beta-marketing-proyecto-fase-beta-marketing/src/types.ts
export interface Campaign {
  id: string;
  name: string;
  brand?: string;
  image_url?: string;
  channel: 'Email' | 'Social' | 'Display' | 'Multi';
  status: 'Activa' | 'Pausada' | 'Completada' | 'Borrador' | 'Pendiente de Pago' | 'Aprobada';
  estado_moderacion?: 'pendiente' | 'aprobada' | 'necesita_revision' | string;
  leads: number;
  reach: string;
  ctr: number;
  startDate: string;
  endDate?: string;
  descripcion?: string;
  objetivo?: string;
  presupuesto?: number;
  fechaInicio?: string;
  fechaFin?: string;
  id_cliente?: string;
  id_usuario?: string;
  creatorName?: string;
  creatorRole?: string;
}
// ...
export interface Pago {
  id_pago: string;
  id_campana?: string;
  monto: number;
  itbis?: number;
  total_con_itbis?: number;
  ncf?: string;
  estado_dgii: 'Pendiente' | 'Aceptado' | 'Rechazado';
  metodo_pago: 'PayPal' | 'Transferencia' | 'Efectivo' | 'Tarjeta';
  rnc_cedula?: string;
  razon_social?: string;
  tipo_comprobante?: string;
  fecha: string;
  nombre_campana?: string;
  url_dgii?: string;
}